requirements.txt
flask==2.0.1
werkzeug==2.0.1
numpy==1.21.0
torch==1.9.0+cpu
torchvision==0.10.0+cpu
faiss-cpu==1.7.1
scikit-image==0.18.2
pillow==8.3.1
matplotlib==3.4.2
pandas==1.3.0
tqdm==4.61.2 